requirements.txt
playwright>=1.49.0
playwright-stealth>=1.0.6
tenacity>=8.2.0
mcp>=1.0.0
httpx>=0.27.0
readability-lxml>=0.8
lxml>=5.0
